Sampling rate in begin is truncated to 0.
begin() takes a uint8_t but the samplerate is 16bit.
Hence the sample rate gets truncated to 0.

+++ Adafruit_TMP006.cpp
@@ -25,7 +25,7 @@ Adafruit_TMP006::Adafruit_TMP006(uint8_t i2caddr) {
 }
 
 
-boolean Adafruit_TMP006::begin(uint8_t samplerate) {
+boolean Adafruit_TMP006::begin(uint16_t samplerate) {
   Wire.begin();
 
   write16(TMP006_CONFIG, TMP006_CFG_MODEON | TMP006_CFG_DRDYEN | samplerate);
